+++ packages/ui5-test-writer/test/unit/expected-output/freestyle/test_ff_1.120/webapp/test/unit/AllTests.js
@@ -1,0 +1,5 @@
+sap.ui.define([
+	"test_ff_1/120/test/unit/controller/View1.controller"
+], function () {
+	"use strict";
+});