+++ packages/mockserver-config-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/mockserver-config-writer
 
+## 0.9.9
+
+### Patch Changes
+
+-   Updated dependencies [4e0bd83]
+    -   @sap-ux/project-access@1.30.8
+
 ## 0.9.8
 
 ### Patch Changes