+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.1.53
+
+### Patch Changes
+
+-   @sap-ux/axios-extension@1.22.4
+-   @sap-ux/odata-service-inquirer@2.5.25
+-   @sap-ux/telemetry@0.6.11
+-   @sap-ux/abap-deploy-config-inquirer@1.5.2
+-   @sap-ux/inquirer-common@0.7.28
+
+## 0.1.52
+
+### Patch Changes
+
+-   Updated dependencies [2cae662]
+    -   @sap-ux/odata-service-inquirer@2.5.24
+
+## 0.1.51
+
+### Patch Changes
+
+-   Updated dependencies [ffac61c]
+    -   @sap-ux/odata-service-inquirer@2.5.23
+    -   @sap-ux/axios-extension@1.22.4
+    -   @sap-ux/abap-deploy-config-inquirer@1.5.1
+
+## 0.1.50
+
+### Patch Changes
+
+-   Updated dependencies [089b56f]
+    -   @sap-ux/abap-deploy-config-inquirer@1.5.0
+
+## 0.1.49
+
+### Patch Changes
+
+-   Updated dependencies [45bbe6c]
+    -   @sap-ux/odata-service-inquirer@2.5.22
+
+## 0.1.48
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.38
+-   @sap-ux/inquirer-common@0.7.27
+-   @sap-ux/odata-service-inquirer@2.5.21
+
 ## 0.1.47
 
 ### Patch Changes